requirements.txt
attr==0.3.2
ConfigParser==7.2.0
contextlib2==21.6.0
cryptography==45.0.4
Cython==3.0.12
dl==0.1.0
docutils==0.21.2
filelock==3.18.0
Flask==3.1.1
googletrans==4.0.2
gtts==2.5.4
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.5
Jinja2==3.1.6
jnius==1.1.0
keyring==25.6.0
langchain_community==0.3.26
mock==5.2.0
numpy==2.3.1
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
Pillow==11.2.1
protobuf==6.31.1
pyOpenSSL==25.1.0
railroad==0.5.0
redis==5.2.1
Sphinx==8.2.3
thread==2.0.5
tornado==6.4.2
trove_classifiers==2025.5.9.12
ultralytics==8.3.158
urllib3_secure_extra==0.1.0
Werkzeug==3.1.3
xmlrpclib==1.0.1
